Move the "trace" logging code to a dedicated module.
